fix empty assets bug

+++ server/adaptors/integrations/integrations_kibana_backend.ts
@@ -19,14 +19,16 @@ export class IntegrationsKibanaBackend implements IntegrationsAdaptor {
 
   deleteIntegrationInstance = async (id: string): Promise<any> => {
     const children: any = await this.client.get('integration-instance', id);
-    children.attributes.assets
-      .map((i) => {
-        return { id: i.assetId, type: i.assetType };
-      })
-      .forEach(async (element) => {
-        await this.client.delete(element.type, element.id);
-      });
-    const result = await this.client.delete('integration-instance', id);
+    if (children.attributes.assets) {
+      children.attributes.assets
+        .map((i) => {
+          return { id: i.assetId, type: i.assetType };
+        })
+        .forEach(async (element) => {
+          await this.client.delete(element.type, element.id);
+        });
+      const result = await this.client.delete('integration-instance', id);
+    }
     return Promise.resolve(result);
   };
